buzz

+++ ProgrameringsProjekt/ArcadeGame/Inc/buzz.h
@@ -5,7 +5,7 @@
 #include "stm32f30x_conf.h" // STM32 config
 #include "30010_io.h" 		// Input/output library for this course
 
-void buzz(int16_t lyd);
+void buzz(int32_t lyd);
